% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tyeRman.R
\name{get_summary_stats}
\alias{get_summary_stats}
\title{get_summary_stats
This function is a custom add-on for use with pairs command.}
\usage{
get_summary_stats(x)
}
\arguments{
\item{x}{is vector}
}
\description{
get_summary_stats
This function is a custom add-on for use with pairs command.
}
\examples{
get_summary_stats(x)
}
\keyword{statistics}

