% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tyeRman.R
\name{impute}
\alias{impute}
\title{impute
This function is a quick and dirty imputation function}
\usage{
impute(d, func)
}
\arguments{
\item{d}{is a dataframe containg columns with NA to be imputed, func is the function to sue to calc the imputed value, uses median currently.}
}
\description{
impute
This function is a quick and dirty imputation function
}
\examples{
impute(d, "median")
}
\keyword{impute}

