% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tyeRman.R
\name{panel_cor}
\alias{panel_cor}
\title{panel.cor
This function is a custom add-on for use with pairs command.}
\usage{
panel_cor(x, y, digits = 2, prefix = "", cex.cor, ...)
}
\arguments{
\item{x, }{y are numeric vectors, digits is number of sig figs to display}
}
\description{
panel.cor
This function is a custom add-on for use with pairs command.
}
\examples{
panel.cor()
}
\keyword{graphics}

